requirements.txt
python-dotenv~=0.15.0
requests~=2.26.0
prometheus_client~=0.11.0